+++ src/routes/(app)/profile/_components/ProfileHeader.svelte
@@ -1,16 +1,18 @@
 <script lang="ts">
 	import Avatar from '$lib/components/Avatar.svelte';
-	import { user } from '$lib/stores/user';
+	import { page } from '$app/stores';
+
+	const user = $derived($page.data.user);
 </script>
 
-{#if $user}
+{#if user}
 	<section>
 		<!-- TODO Set default picture in db -->
-		<Avatar src={$user.picture ?? ''} alt={$user.username ?? 'User avatar'} />
+		<Avatar src={user.picture ?? ''} alt={user.username ?? 'User avatar'} />
 		<div class="flex flex-col">
-			<p class="text-xl font-bold">{$user.username}</p>
+			<p class="text-xl font-bold">{user.username}</p>
 			<p class="text-sm text-muted-foreground">
-				Member since {$user.createdAt.toLocaleDateString()}
+				Member since {user.createdAt.toLocaleDateString()}
 			</p>
 		</div>
 	</section>