+++ src/routes/(app)/submissions/+page.server.ts
@@ -2,6 +2,7 @@ import { db } from '$lib/server/db';
 import { count } from 'drizzle-orm/sql/functions';
 import type { PageServerLoad } from './$types';
 import submissions from '$lib/server/db/schema/submission';
+import { eq, or } from 'drizzle-orm';
 
 // TODO Handling invalid page numbers
 // TODO Combining pagination with filtering/sorting
@@ -9,16 +10,20 @@ import submissions from '$lib/server/db/schema/submission';
 export const load = (async ({ locals, url }) => {
 	const page = Number(url.searchParams.get('page') ?? '1');
 	const sortBy = url.searchParams.get('sortBy') ?? 'date_asc';
-	const pageSize = 6;
+	const pageSize = 30;
 
 	const result = await db.transaction(async (tx) => {
 		const result = await tx.query.submissions.findMany({
 			with: {
-				user: true,
+				user: { columns: { id: true, username: true, picture: true } },
 				media: true,
 				reactions: true,
 				category: true
 			},
+			where: (submissions, { or, eq }) =>
+				locals.user
+					? or(eq(submissions.userId, locals.user.id), eq(submissions.status, 'published'))
+					: eq(submissions.status, 'published'),
 			orderBy: (items, { asc, desc }) => {
 				switch (sortBy) {
 					case 'date_asc':
@@ -33,13 +38,18 @@ export const load = (async ({ locals, url }) => {
 			offset: (page - 1) * pageSize
 		});
 
-		const [totalCount]: {
-			count: number;
-		}[] = await tx.select({ count: count() }).from(submissions);
+		const [totalCount] = await tx
+			.select({ count: count() })
+			.from(submissions)
+			.where(
+				locals.user
+					? or(eq(submissions.userId, locals.user.id), eq(submissions.status, 'published'))
+					: eq(submissions.status, 'published')
+			);
 
 		return { submissions: result, totalCount };
 	});
-	console.log(result);
+
 	const categories = await db.query.categories.findMany({
 		with: { categoriesToEvents: { with: { event: { columns: { name: true, id: true } } } } }
 	});